requirements.txt
chardet==4.0.0
et-xmlfile==1.1.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.5
pysrt==1.1.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
vtt-to-srt3==0.1.9.1
